refactor scatterpolargl - add calc, plot and hover files
