+++ src/traces/scatterpolargl/hover.js
@@ -1,0 +1,48 @@
+/**
+* Copyright 2012-2019, Plotly, Inc.
+* All rights reserved.
+*
+* This source code is licensed under the MIT license found in the
+* LICENSE file in the root directory of this source tree.
+*/
+
+'use strict';
+
+var hover = require('../scattergl/hover');
+var makeHoverPointText = require('../scatterpolar/hover').makeHoverPointText;
+
+function hoverPoints(pointData, xval, yval, hovermode) {
+    var cd = pointData.cd;
+    var stash = cd[0].t;
+    var rArray = stash.r;
+    var thetaArray = stash.theta;
+
+    var scatterPointData = hover.hoverPoints(pointData, xval, yval, hovermode);
+    if(!scatterPointData || scatterPointData[0].index === false) return;
+
+    var newPointData = scatterPointData[0];
+
+    if(newPointData.index === undefined) {
+        return scatterPointData;
+    }
+
+    var subplot = pointData.subplot;
+    var cdi = newPointData.cd[newPointData.index];
+    var trace = newPointData.trace;
+
+    // augment pointData with r/theta param
+    cdi.r = rArray[newPointData.index];
+    cdi.theta = thetaArray[newPointData.index];
+
+    if(!subplot.isPtInside(cdi)) return;
+
+    newPointData.xLabelVal = undefined;
+    newPointData.yLabelVal = undefined;
+    makeHoverPointText(cdi, trace, subplot, newPointData);
+
+    return scatterPointData;
+}
+
+module.exports = {
+    hoverPoints: hoverPoints
+};
